Rename core package to client and hide peerPresenceMap from attachment
