+++ src/client/auth.ts
